# backend/openapi.json
{
  "components": {
    "schemas": {
      "BloodType": {
        "enum": [
          "A+",
          "A-",
          "B+",
          "B-",
          "AB+",
          "AB-",
          "O+",
          "O-"
        ],
        "title": "BloodType",
        "type": "string"
      },
      "ErrorResponse": {
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "errors": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "title": "Errors",
            "type": "array"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
      "FieldError": {
        "properties": {
          "field": {
            "title": "Field",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "field",
          "message"
        ],
        "title": "FieldError",
        "type": "object"
      },
...
      "NoteCreate": {
        "properties": {
          "content": {
            "maxLength": 5000,
            "minLength": 1,
            "title": "Content",
            "type": "string"
          },
          "noted_at": {
            "description": "When the observation was made.",
            "format": "date-time",
            "title": "Noted At",
            "type": "string"
          }
        },
        "required": [
          "content",
          "noted_at"
        ],
        "title": "NoteCreate",
        "type": "object"
      },
      "NoteListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/NoteRead"
            },
            "title": "Items",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total"
        ],
        "title": "NoteListResponse",
        "type": "object"
      },
      "NoteRead": {
        "properties": {
          "content": {
            "title": "Content",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "noted_at": {
            "format": "date-time",
            "title": "Noted At",
            "type": "string"
          },
          "patient_id": {
            "format": "uuid",
            "title": "Patient Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "patient_id",
          "content",
          "noted_at",
          "created_at"
        ],
        "title": "NoteRead",
        "type": "object"
      },
...
      "PatientRead": {
        "properties": {
          "address_line1": {
            "title": "Address Line1",
            "type": "string"
          },
          "address_line2": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address Line2"
          },
          "age": {
            "readOnly": true,
            "title": "Age",
            "type": "integer"
          },
          "allergies": {
            "items": {
              "type": "string"
            },
            "title": "Allergies",
            "type": "array"
          },
          "blood_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BloodType"
              },
              {
                "type": "null"
              }
            ]
          },
          "city": {
            "title": "City",
            "type": "string"
          },
          "conditions": {
            "items": {
              "type": "string"
            },
            "title": "Conditions",
            "type": "array"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "date_of_birth": {
            "format": "date",
            "title": "Date Of Birth",
            "type": "string"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "first_name": {
            "title": "First Name",
            "type": "string"
          },
          "full_name": {
            "readOnly": true,
            "title": "Full Name",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "last_name": {
            "title": "Last Name",
            "type": "string"
          },
          "last_visit_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Visit At"
          },
          "phone": {
            "title": "Phone",
            "type": "string"
          },
          "postal_code": {
            "title": "Postal Code",
            "type": "string"
          },
          "state": {
            "title": "State",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/PatientStatus"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "first_name",
          "last_name",
          "date_of_birth",
          "email",
          "phone",
          "address_line1",
          "address_line2",
          "city",
          "state",
          "postal_code",
          "blood_type",
          "status",
          "allergies",
          "conditions",
          "last_visit_at",
          "created_at",
          "updated_at",
          "full_name",
          "age"
        ],
        "title": "PatientRead",
        "type": "object"
      },
...
      "PatientStatus": {
        "enum": [
          "active",
          "pending",
          "inactive",
          "discharged"
        ],
        "title": "PatientStatus",
        "type": "string"
      },
      "PatientSummary": {
        "properties": {
          "clinical": {
            "$ref": "#/components/schemas/SummaryClinical"
          },
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          },
          "generated_by": {
            "title": "Generated By",
            "type": "string"
          },
          "identifiers": {
            "$ref": "#/components/schemas/SummaryIdentifiers"
          },
          "narrative": {
            "title": "Narrative",
            "type": "string"
          },
          "patient_id": {
            "format": "uuid",
            "title": "Patient Id",
            "type": "string"
          }
        },
        "required": [
          "patient_id",
          "identifiers",
          "clinical",
          "narrative",
          "generated_by",
          "generated_at"
        ],
        "title": "PatientSummary",
        "type": "object"
      },
      "PatientUpdate": {
        "properties": {
          "address_line1": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Address Line1",
            "type": "string"
          },
          "address_line2": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address Line2"
          },
          "allergies": {
            "default": [],
            "items": {
              "maxLength": 120,
              "minLength": 1,
              "type": "string"
            },
            "maxItems": 50,
            "title": "Allergies",
            "type": "array"
          },
          "blood_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BloodType"
              },
              {
                "type": "null"
              }
            ]
          },
          "city": {
            "maxLength": 100,
            "minLength": 1,
            "title": "City",
            "type": "string"
          },
          "conditions": {
            "default": [],
            "items": {
              "maxLength": 120,
              "minLength": 1,
              "type": "string"
            },
            "maxItems": 50,
            "title": "Conditions",
            "type": "array"
          },
          "date_of_birth": {
            "format": "date",
            "title": "Date Of Birth",
            "type": "string"
          },
          "email": {
            "anyOf": [
              {
                "format": "email",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "first_name": {
            "maxLength": 100,
            "minLength": 1,
            "title": "First Name",
            "type": "string"
          },
          "last_name": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Last Name",
            "type": "string"
          },
          "last_visit_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Visit At"
          },
          "phone": {
            "maxLength": 30,
            "minLength": 7,
            "title": "Phone",
            "type": "string"
          },
          "postal_code": {
            "maxLength": 20,
            "minLength": 3,
            "title": "Postal Code",
            "type": "string"
          },
          "state": {
            "maxLength": 50,
            "minLength": 2,
            "title": "State",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/PatientStatus",
            "default": "active"
          }
        },
        "required": [
          "first_name",
          "last_name",
          "date_of_birth",
          "phone",
          "address_line1",
          "city",
          "state",
          "postal_code"
        ],
        "title": "PatientUpdate",
        "type": "object"
      },
...
      "SummaryClinical": {
        "properties": {
          "allergies": {
            "items": {
              "type": "string"
            },
            "title": "Allergies",
            "type": "array"
          },
          "conditions": {
            "items": {
              "type": "string"
            },
            "title": "Conditions",
            "type": "array"
          },
          "last_visit_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Visit At"
          },
          "note_count": {
            "title": "Note Count",
            "type": "integer"
          }
        },
        "required": [
          "conditions",
          "allergies",
          "last_visit_at",
          "note_count"
        ],
        "title": "SummaryClinical",
        "type": "object"
      },
      "SummaryIdentifiers": {
        "properties": {
          "age": {
            "title": "Age",
            "type": "integer"
          },
          "blood_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BloodType"
              },
              {
                "type": "null"
              }
            ]
          },
          "date_of_birth": {
            "title": "Date Of Birth",
            "type": "string"
          },
          "full_name": {
            "title": "Full Name",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/PatientStatus"
          }
        },
        "required": [
          "full_name",
          "age",
          "date_of_birth",
          "blood_type",
          "status"
        ],
        "title": "SummaryIdentifiers",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Patient management API for a medical practice.",
    "title": "Healthcare Dashboard API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/patients/{patient_id}": {
      "delete": {
        "operationId": "delete_patient_patients__patient_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "patient_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Patient Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Patient not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation failed"
          }
        },
        "summary": "Delete a patient and their notes",
        "tags": [
          "patients"
        ]
      },
      "get": {
        "operationId": "read_patient_patients__patient_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "patient_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Patient Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Patient not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation failed"
          }
        },
        "summary": "Get a patient",
        "tags": [
          "patients"
        ]
      },
      "put": {
        "operationId": "update_patient_patients__patient_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "patient_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Patient Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatientUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Patient not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation failed"
          }
        },
        "summary": "Replace a patient",
        "tags": [
          "patients"
        ]
      }
    },
    "/patients/{patient_id}/notes": {
      "get": {
        "operationId": "list_notes_patients__patient_id__notes_get",
        "parameters": [
          {
            "in": "path",
            "name": "patient_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Patient Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoteListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Patient or note not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation failed"
          }
        },
        "summary": "List a patient's notes",
        "tags": [
          "notes"
        ]
      },
      "post": {
        "operationId": "create_note_patients__patient_id__notes_post",
        "parameters": [
          {
            "in": "path",
            "name": "patient_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Patient Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NoteCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoteRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Patient or note not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation failed"
          }
        },
        "summary": "Add a clinical note",
        "tags": [
          "notes"
        ]
      }
    },
    "/patients/{patient_id}/notes/{note_id}": {
      "delete": {
        "operationId": "delete_note_patients__patient_id__notes__note_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "patient_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Patient Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "note_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Note Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Patient or note not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation failed"
          }
        },
        "summary": "Delete a note",
        "tags": [
          "notes"
        ]
      }
    },
    "/patients/{patient_id}/summary": {
      "get": {
        "operationId": "read_summary_patients__patient_id__summary_get",
        "parameters": [
          {
            "in": "path",
            "name": "patient_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Patient Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Patient or note not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation failed"
          }
        },
        "summary": "Human-readable summary of the profile and notes",
        "tags": [
          "notes"
        ]
      }
    }
  }
}